Refactor options normalization
