+++ src/main/normalize-options.js
